+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 3.1.0
-- Add `FreError.getError()` params are now optional
+- `FreError.getError()` params are now optional
 - Add `FREObject.className`
 - Add `FreObjectSwift.className`
 - Add `FreObjectSwift.toString()`
@@ -7,12 +7,11 @@
 - Add `[Date]` support to `FREArray`
 - Add `[NSNumber]` support to `FREArray`
 - Declare `FREArray` as open
-- Change: `FREArray` now uses Vector.<Type> throughout
+- Change: `FREArray` now uses `Vector.<Type>` throughout
 - Rename `FreObjectTypeSwift.cls` to `FreObjectTypeSwift.class`
 - Improve `ANEUtils.map` performance
 - Obsolete `FREObject.setProp()`
 - Obsolete `FREObject.getProp()`
-- Obsolete `sendEvent`
 - Fix spelling mistakes in docs
 
 ### 3.0.0